requirements.txt
asgiref==3.8.1
cfgv==3.4.0
coverage==7.6.10
distlib==0.3.9
Django==5.1.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
whitenoise
filelock==3.16.1
identify==2.6.4
inflection==0.5.1
iniconfig==2.0.0
nodeenv==1.9.1
numpy==2.1.3
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
slack_sdk==3.34.0
sqlparse==0.5.3
uritemplate==4.1.1
virtualenv==20.28.0
